+++ tests/ecosystem/upgrade/test_noobaa.py
@@ -7,10 +7,10 @@ from ocs_ci.framework.pytest_customization.marks import (
 )
 from ocs_ci.ocs import constants
 from ocs_ci.ocs.constants import BS_OPTIMAL
-from tests.helpers import (
-    retrieve_test_objects_to_pod, sync_object_directory
+from ocs_ci.ocs.bucket_utils import (
+    retrieve_test_objects_to_pod, sync_object_directory,
+    verify_s3_object_integrity
 )
-from tests.helpers import verify_s3_object_integrity
 
 logger = logging.getLogger(__name__)
