+++ tests/manage/mcg/test_object_integrity.py
@@ -4,12 +4,11 @@ import pytest
 
 from ocs_ci.framework.testlib import ManageTest, tier1, tier2, tier3
 from ocs_ci.ocs import constants
-from tests.helpers import (
+from ocs_ci.ocs.bucket_utils import (
     retrieve_test_objects_to_pod, sync_object_directory,
-    craft_s3_command
+    craft_s3_command, verify_s3_object_integrity,
+    retrieve_anon_s3_resource
 )
-from tests.helpers import retrieve_anon_s3_resource
-from tests.helpers import verify_s3_object_integrity
 
 logger = logging.getLogger(__name__)
