+++ tests/manage/rgw/test_object_integrity.py
@@ -1,7 +1,7 @@
 import logging
 
 import pytest
-from tests.helpers import (
+from ocs_ci.ocs.bucket_utils import (
     retrieve_test_objects_to_pod, sync_object_directory,
     verify_s3_object_integrity
 )
